#16 - implemented importing recipes and vmtemplates: step1

+++ ecloud-manager-web/src/main/java/org/ecloudmanager/web/faces/RecipesController.java
@@ -25,39 +25,53 @@
 package org.ecloudmanager.web.faces;
 
 import com.rits.cloning.Cloner;
+import org.apache.commons.lang3.StringUtils;
+import org.bson.types.ObjectId;
+import org.ecloudmanager.deployment.core.Deployable;
 import org.ecloudmanager.deployment.vm.provisioning.Recipe;
+import org.ecloudmanager.jeecore.web.faces.Controller;
 import org.ecloudmanager.jeecore.web.faces.FacesSupport;
 import org.ecloudmanager.repository.template.RecipeRepository;
 import org.ecloudmanager.service.template.RecipeService;
+import org.ecloudmanager.service.template.VirtualMachineTemplateService;
 import org.primefaces.event.CloseEvent;
 import org.primefaces.event.SelectEvent;
 import org.primefaces.event.UnselectEvent;
 
 import javax.annotation.PostConstruct;
-import javax.faces.bean.ManagedBean;
-import javax.faces.bean.ManagedProperty;
-import javax.faces.bean.SessionScoped;
+import javax.faces.application.FacesMessage;
 import javax.inject.Inject;
 import java.io.Serializable;
 import java.util.List;
 
-@SessionScoped
-@ManagedBean
+@Controller
 public class RecipesController extends FacesSupport implements Serializable {
-    public final static String NEW_RECIPE_ID = "(new)";
     private static final long serialVersionUID = -7592006943742318320L;
     @Inject
     private transient RecipeRepository recipeRepository;
     @Inject
     private transient RecipeService recipeService;
     @Inject
+    private transient VirtualMachineTemplateService virtualMachineTemplateService;
+    @Inject
     private transient Cloner cloner;
-    @ManagedProperty(value = "#{recipeController}")
+    @Inject
     private transient RecipeController recipeController;
+    @Inject
+    private ApplicationDeploymentEditorController applicationDeploymentEditorController;
+
     private List<Recipe> recipes;
 
     public void setRecipeController(RecipeController recipeController) {
         this.recipeController = recipeController;
+    }
+
+    public ApplicationDeploymentEditorController getApplicationDeploymentEditorController() {
+        return applicationDeploymentEditorController;
+    }
+
+    public void setApplicationDeploymentEditorController(ApplicationDeploymentEditorController applicationDeploymentEditorController) {
+        this.applicationDeploymentEditorController = applicationDeploymentEditorController;
     }
 
     @PostConstruct
@@ -65,8 +79,10 @@ public class RecipesController extends FacesSupport implements Serializable {
         refresh();
     }
 
-    private void refresh() {
-        recipes = recipeRepository.getAll();
+    public void refresh() {
+        Deployable owner = applicationDeploymentEditorController == null || applicationDeploymentEditorController.getDeployment() == null ?
+                null : applicationDeploymentEditorController.getDeployment();
+        recipes = recipeRepository.getAll(owner);
     }
 
 
@@ -84,6 +100,8 @@ public class RecipesController extends FacesSupport implements Serializable {
         } else {
             recipeService.update(value);
         }
+
+        refresh();
     }
 
     public void cancel() {
@@ -91,7 +109,7 @@ public class RecipesController extends FacesSupport implements Serializable {
         if (value.isNew()) {
             recipes.remove(value);
         } else {
-            Recipe reloaded = recipeRepository.get(value.getOldId());
+            Recipe reloaded = recipeRepository.get(value.getId());
             int idx = recipes.indexOf(value);
             recipes.set(idx, reloaded);
         }
@@ -116,19 +134,56 @@ public class RecipesController extends FacesSupport implements Serializable {
 
 
     public Recipe getNewRecipe() {
-        recipeController.setValue(new Recipe(NEW_RECIPE_ID));
+        Recipe recipe = new Recipe();
+        recipe.setId(new ObjectId());
+        updateOwner(recipe);
+        recipeController.setValue(recipe);
         return recipeController.getValue();
+    }
+
+    private void updateOwner(Recipe recipe) {
+        if (
+                applicationDeploymentEditorController != null &&
+                applicationDeploymentEditorController.getDeployment() != null
+        ) {
+            recipe.setOwner(applicationDeploymentEditorController.getDeployment());
+        } else {
+            recipe.setOwner(null);
+        }
     }
 
     public void deleteSelected() {
         if (recipeController.getValue() != null) {
-            recipeService.remove(recipeController.getValue());
+            List<String> usages = virtualMachineTemplateService.findRecipeUsages(recipeController.getValue());
+            if (usages.isEmpty()) {
+                recipeService.remove(recipeController.getValue());
+                recipeController.setValue(null);
+            } else {
+                addMessage(new FacesMessage(
+                        FacesMessage.SEVERITY_ERROR,
+                        "Cannot delete the selected recipe",
+                        "Cannot delete the selected recipe. It is used by the following objects: <br>" + StringUtils.join(usages, "<br>"))
+                );
+            }
         }
-        recipeController.setValue(null);
+        refresh();
     }
 
 
     public void handleClose(CloseEvent event) {
         event.getSource();
+    }
+
+    public void onImportRecipeReturn(SelectEvent event) {
+        List<Recipe> newRecipes = (List<Recipe>) event.getObject();
+        if (newRecipes != null) {
+            newRecipes.stream().forEach(recipe -> {
+                Recipe newRecipe = cloner.deepClone(recipe);
+                newRecipe.setId(new ObjectId());
+                updateOwner(newRecipe);
+                recipeService.saveWithUniqueName(newRecipe);
+                recipes.add(newRecipe);
+            });
+        }
     }
 }
